CATSSI-383: force admin to set correct hostname
avoid using default value from hostname - which is salt-minion ID
and might not fit to use as outgoing address.

E.g: we are set minionID/hostname: uba-mail, it will by default
set outgoing address to uba-mail.unblockapp.com. But we want
mail.unblockapp.com. So set a pillar here to force that.